Now showing preloader when scanning for updates. Showing message after adding multiple directories.
