+++ GitMon.njsproj
@@ -40,6 +40,9 @@
     <Compile Include="renderer\badgeHandler.js" />
     <Compile Include="renderer\directoryPickerCaller.js" />
     <Compile Include="renderer\externalLinkHandler.js" />
+    <Compile Include="gitHandler.js">
+      <SubType>Code</SubType>
+    </Compile>
     <Compile Include="renderer\gitStatusResultDOMHandler.js" />
     <Compile Include="renderer\mainDivSizeHandler.js" />
     <Compile Include="renderer\preloader.js" />
